Se corrigen Bugs y se agrega totalizador de items en la opción de descarga de XLSX

+++ ClienteServicio/Controllers/ClienteController.cs
@@ -1,4 +1,6 @@
-﻿using ClienteServicio.Models;
+﻿using ClienteServicio.helpers;
+using ClienteServicio.mappers;
+using ClienteServicio.Models;
 using ClienteServicio.Repository;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
@@ -24,49 +26,70 @@ namespace ClienteServicio.Controllers
         [HttpGet("all")]
         public IActionResult GetAllCustomer()
         {
-            List<Customer> data = new List<Customer>();
-            data = _customerRepository.GetAllCustomers();
-            if (!data.Any())
+            try
             {
-                return NotFound();
+                List<CustomerViewModel> data = new List<CustomerViewModel>();
+                var result = _customerRepository.GetAllCustomers();
+                if (result.Count == 0)
+                {
+                    return NotFound(new { statusCode = 404, message = "No customers found." });
+                }
+
+                return Ok(Commons.ConvertDtoListToEntityList<CustomerViewModel, Customer>(result));
             }
-            return Ok(data);
+            catch (Exception ex)
+            {
+                return StatusCode(Commons.GetStatusCodeFromException(ex), new { message = ex.Message });
+            }
+           
         }
 
-        [HttpPut("{rut}/name")]
-        public IActionResult UpdateName(int rut, [FromBody] Customer newCustomer)
+        [HttpPut("name")]
+        public IActionResult UpdateName([FromBody] CustomerViewModel newCustomer)
         {
-            var customer = _customerRepository.GetCustomerByRut(newCustomer.rut, newCustomer.dv);
-            if (customer == null)
+            try
             {
-                return NotFound(new { message = "Customer not found." });
+                if (newCustomer == null || (newCustomer.rut <= 0 && string.IsNullOrEmpty(newCustomer.dv))) {
+                    return BadRequest(new { statusCode = 400, message = "Invalid input data." });
+                }
+                var customer = _customerRepository.GetCustomerByRut(newCustomer.rut, newCustomer.dv);
+                if (customer == null)
+                {
+                    return NotFound(new { message = "Customer not found." });
+                }
+
+                customer.customer = newCustomer.customer;
+                _customerRepository.Update(customer); // Actualiza en la base de datos
+
+
+                return Ok(new { message = "Customer updated successfully." });
             }
-
-            customer.customer = newCustomer.customer;
-            _customerRepository.Update(customer); // Actualiza en la base de datos
-
-
-            return Ok(new { message = "Customer updated successfully." });
+            catch (Exception ex)
+            {
+                return StatusCode(Commons.GetStatusCodeFromException(ex), new { message = ex.Message });
+            }
+            
         }
 
 
         [HttpPost("add")]
-        public IActionResult AddCustomer([FromBody] Customer newCustomer)
+        public IActionResult AddCustomer([FromBody] CustomerViewModel newCustomer)
         {
-            if (newCustomer == null)
-            {
-                return BadRequest("El cliente no puede ser nulo.");
-            }
-
             try
             {
-                _customerRepository.Add(newCustomer);
+                if (newCustomer == null)
+                {
+                    return BadRequest(new { statusCode = 400, message = "Invalid input data." });
+                }
+
+                var customer = Commons.ConvertDtoToEntity<Customer,CustomerViewModel>(newCustomer);
+                _customerRepository.Add(customer);
                 return CreatedAtAction(nameof(AddCustomer), new { id = newCustomer.rut }, newCustomer);
             }
             catch (Exception ex)
             {
                 // Manejo de errores
-                return StatusCode(500, $"Error al agregar el cliente: {ex.Message}");
+                return StatusCode(Commons.GetStatusCodeFromException(ex), new { message = ex.Message });
             }
         }
 
@@ -74,15 +97,28 @@ namespace ClienteServicio.Controllers
         [HttpDelete("{id}/{dv}/down")]
         public IActionResult Remove(int id, string dv)
         {
-            var cust = _customerRepository.GetCustomerByRut(id, dv);
-            if (cust == null)
+            try
             {
-                return NotFound(new { message = "Customer not found." });
-            }
+                if (id <= 0 || string.IsNullOrEmpty(dv))
+                {
+                    return BadRequest(new { statusCode = 400, message = "Invalid input data." });
+                }
+                var cust = _customerRepository.GetCustomerByRut(id, dv);
+                if (cust == null)
+                {
+                    return NotFound(new { message = "Customer not found." });
+                }
+                _customerRepository.Delete(cust);
 
-            _customerRepository.Delete(cust);
+                return Ok(new { message = "Customer removed successfully." });
+            }
+            catch(Exception ex)
+            {
+                return StatusCode(Commons.GetStatusCodeFromException(ex), new { message = ex.Message });
+            }
             
-            return Ok(new { message = "Customer removed successfully." });
+
+            
         }
     }
 }
